feat: Add description field and update default reference format
This commit introduces a `description` field to the `create_transaction`
method in the `SavingsTransactionService`.  This allows for storing
additional context related to the transaction.

The default transaction reference format has also been updated from
`{transaction_type}_{uuid}` to `TXN_{transaction_type}_{uuid}` for
better identification and consistency.  This makes it easier to
distinguish savings transactions from other types of transactions.