feat: Register LocalAuthPlugin for Windows platform
This commit adds the `LocalAuthPlugin` to the list of registered
plugins for the Windows platform in `generated_plugin_registrant.cc`.
This enables local authentication functionalities (e.g., fingerprint
scanning, PIN authentication) to be used within the Flutter application
when running on Windows. The `LocalAuthPluginRegisterWithRegistrar`
function is called to register the plugin with the Flutter plugin
registry.