feat: Add Firebase and local notification dependencies
This commit adds the following dependencies to the `pubspec.yaml` file:

- `firebase_core`: For core Firebase functionality.
- `flutter_local_notifications`: For displaying local notifications.
- `firebase_messaging`: For handling Firebase Cloud Messaging (FCM) push notifications.
- `permission_handler`: For requesting necessary permissions for FCM.

These dependencies are added to enable push notification functionality within the application, leveraging Firebase Cloud Messaging. The `permission_handler` is required to request permissions needed by the mobile OS (Android/iOS) for the notifications to be received and displayed.