fix: Update iOS platform version to 13.0 in Podfile
The minimum iOS deployment target in the Podfile has been updated
from 12.0 to 13.0. This change ensures compatibility with newer
features and dependencies used in the project and aligns with the
recommended deployment target for the current app features.  This
change was needed because some dependencies now require a minimum
iOS version of 13.